MAJOR: Introduced detect open from Widget Widgets on macOS can now open the app and the app detects this.
To handle the app being open and widgets being on desktop what occurs is every time its treated as a "new open", aka the inital URL is pushed over, and its treated like new open.

Its done this way to ensure stable running. This also mimics normal mac app opening behavior.

For example, when you have 2 reminder widgets set to two different reminders, then what occurs is the app switches to that version of the open reminder.

With this in mind I've implemented it so such occur. When you press on a widget in the example simple program, the app always re shows the appeared from address, beacause the app in essenece is releaunced from there.

# What changed:
- App group for macOS so far HAS to follow the layout of`Team.org.name` rather than `group.org.name`

# What Else is New:

- Improved stability with sending updates to widgets. Lots of problems resided in needed to delete derived data and early executable builds of the app in `/Users/devingrischow/Documents/Code/Plugins/Flutter/home_widget_iterations/mac_integration_prototyping/home_widget/examples/simple_apple_widget/build/macos/Build/Products/Debug`. Simply deleting the app file and having runtime rebuild it fixes alot of the stability issues.

# Warning Debug Features in this Commit.

For debugging opening the app without a debugger, an `app_log` was used. Its located at:

- `/Users/devingrischow/Library/Containers/com.devg.simpleAppleWidget/Data/Documents`

Lots of actions are logged there for testing this early build/prototype.

# Whats Next
- Test Stability of Open From Widget
- Improve stability and connecting to mac app and widgets
- Analytics
- Handling Interactions
- Handle Configuring widgets